refactor: file structure changed to be by domain
