+++ web/src/domains/users/views/CreateAccountView.vue
@@ -40,7 +40,7 @@
 
 <script>
 import { initUserStore } from '@/stores/user'
-import { client } from '../../api-client/client'
+import { client } from '../../../../api-client/client'
 
 export default {
   name: 'CreateAccountView',
