+++ web/src/domains/users/views/DashboardView.vue
@@ -52,9 +52,9 @@
     </v-row>
 </template>
 <script>
-import { initUserStore } from '../../stores/user'
-import { client } from '../../../api-client/client'
-import ShiftTableRow from '../../components/shifts/ShiftTableRow.vue'
+import { initUserStore } from '@/stores/user'
+import { client } from '../../../../api-client/client'
+import ShiftTableRow from '../../shifts/components/ShiftTableRow.vue'
 
 export default {
     name: 'DashboardView',
