+++ web/src/domains/users/views/LoginView.vue
@@ -30,8 +30,8 @@
 </template>
 
 <script>
-import { client } from '../../api-client/client'
-import { initUserStore } from '../stores/user'
+import { client } from '../../../../api-client/client'
+import { initUserStore } from '../../../stores/user'
 
 export default {
   name: 'LoginView',
